Add binary search tree implementation
This commit adds the implementation of a binary search tree in the `trees/binarySearchTree.js` file. The `Tree` class represents the binary search tree, while the `Node` class represents each node in the tree. The `add` method is used to add values to the tree, and the `toObject` method returns the tree as an object. The implementation also includes a test case that creates a binary search tree from an array of numbers.